Add store interface
