+++ server/store/apistore/poll.go
@@ -1,4 +1,4 @@
-package store
+package apistore
 
 import (
 	"errors"
