+++ server/store/apistore/poll_test.go
@@ -1,4 +1,4 @@
-package store
+package apistore
 
 import (
 	"testing"
